+++ storage/database_factory.py
@@ -13,6 +13,7 @@ class DatabaseFactory:
     def get_database_connection(self):
         if self._database is None:
             self._database = MySQLDatabase(self.config.get("Database", "Database"),
+                                           port=int(self.config.get("Database", "Port")),
                                            host=self.config.get("Database", "Host"),
                                            user=self.config.get("Database", "User"),
                                            password=self.config.get("Database", "Password")
@@ -20,6 +21,6 @@ class DatabaseFactory:
 
         return self._database
 
-    # @tasks.loop(seconds=30)
-    # async def stay_alive(self):
-    #     self.get_database_connection().connection().ping(True)
+    @tasks.loop(seconds=30)
+    async def stay_alive(self):
+        self.get_database_connection().connection().ping(True)